use entity manager inside jboss to create test data
- this will ensure second level cache and lucene search index correctness
- less overhead of maintaining another copy of persistence.xml
