+++ functional-test/src/test/java/org/zanata/util/AddUsersRule.java
@@ -2,6 +2,12 @@ package org.zanata.util;
 
 import lombok.extern.slf4j.Slf4j;
 import org.junit.rules.ExternalResource;
+import org.zanata.common.LocaleId;
+import org.zanata.rest.SampleProjectResource;
+
+import com.google.common.collect.Sets;
+
+import static org.zanata.util.SampleProjectClient.checkAndReleaseConnection;
 
 /**
  * @author Patrick Huang <a
@@ -10,17 +16,19 @@ import org.junit.rules.ExternalResource;
 @Slf4j
 public class AddUsersRule extends ExternalResource {
 
-    private SampleProjectProfile profile;
+    private final SampleProjectResource resource = SampleProjectClient.RESOURCE;
 
     @Override
     protected void before() {
-        profile = new SampleProjectProfile();
-        profile.deleteExceptEssentialData();
-        profile.makeSampleUsers();
+        checkAndReleaseConnection(resource.deleteExceptEssentialData());
+        checkAndReleaseConnection(resource.makeSampleUsers());
+        checkAndReleaseConnection(resource.userJoinsLanguageTeams("translator",
+                Sets.newHashSet(LocaleId.FR, new LocaleId("hi"),
+                        new LocaleId("pl"))));
     }
 
     @Override
     protected void after() {
-        profile.deleteExceptEssentialData();
+        checkAndReleaseConnection(resource.deleteExceptEssentialData());
     }
 }
