+++ functional-test/src/test/resources/concordion/org/zanata/feature/dashboard/Dashboard.html
@@ -5,8 +5,8 @@
 </head>
 <body>
 <p>
-  As a user, when I use username: <span c:set="#username">glossarist</span> with
-  password: <span c:set="#password">glossarist</span>
+  As a user, when I use username: <span c:set="#username">admin</span> with
+  password: <span c:set="#password">admin</span>
   to <span c:execute="#result = signInAs(#username, #password)">sign in</span>.
 
   I should redirect to dashboard with <span
