+++ zanata-war/pom.xml
@@ -101,6 +101,27 @@
               <goal>properties</goal>
             </goals>
           </execution>
+          <execution>
+            <!-- install to a temporary location so that functional test can use it -->
+            <id>copy</id>
+            <phase>package</phase>
+            <goals>
+              <goal>copy</goal>
+            </goals>
+            <configuration>
+              <artifactItems>
+                <artifactItem>
+                  <groupId>${project.groupId}</groupId>
+                  <artifactId>${project.artifactId}</artifactId>
+                  <version>${project.version}</version>
+                  <type>war</type>
+                  <overWrite>true</overWrite>
+                  <outputDirectory>${java.io.tmpdir}/zanata-war</outputDirectory>
+                  <destFileName>zanata-test.war</destFileName>
+                </artifactItem>
+              </artifactItems>
+            </configuration>
+          </execution>
         </executions>
         <configuration>
           <failOnWarning>true</failOnWarning>
